requirements.txt
requests==2.32.3
discord.py==2.3.2
